Add 'new' method for the 'pointer' type
vz code currently has a lot of occurrences of this kind of constructs
when creating go wrappers for objc objects:
```
config := &VirtioSoundDeviceConfiguration{
    pointer: pointer{
        ptr: C.newVZVirtioSoundDeviceConfiguration(),
    },
}
```

This is a bit verbose, but most importantly this exposes the 'ptr'
field of the 'pointer' struct. In one of the next commits, I'd like to
rename this field, so as a preparatory commit, this introduces a
`newPointer` method to replace the code above with:
```
config := &VirtioSoundDeviceConfiguration{
    pointer: newPointer(C.newVZVirtioSoundDeviceConfiguration()),
}
```
which is shorter, and hides the `ptr` field from code outside of
objcutil.go.