Make pointer.Ptr() private
Users of Code-Hex/vz  don't need access to the raw objc pointer, but
it's currently exposed as a Ptr() method.
This commit makes this method private (ptr()). This involves renaming
the `ptr` field to `wrappedPtr`, which is easy to after the addition of
newPointer().
This will remove all Ptr() methods from
https://pkg.go.dev/github.com/Code-Hex/vz/v2#BridgedNetworkDeviceAttachment.Ptr