vsock: Use net.FileConn to create VirtioSocketConnection
Currently, newVirtioSocketConnection manually sets its objc file
descriptor to non-blocking, calls dup() on it, ... so that it
matches what go expects from a file descriptor.

go provides a helper doing all of this for us: net.FileConn()
This commit makes use of that helper. This will allow for more code
simplifications in the next commits.

However this removes vz.VirtioSocketConnection.FileDescriptor() since
the file descriptor was created internally by net.FileConn().
It would be possible to keep it, but I'm not sure it's really useful.
It would be more consistent with other go APIs to add a File() method if
we want to expose this.

One side-effect of this change is to fix connections returned by
ConnectToPort. At the moment, they are missing a call to Dup() for the
file descriptor they use. Without this, their file descriptor is closed
too early, and the connection cannot be used.

This fixes https://github.com/Code-Hex/vz/issues/54