macos: Don't expose VZMacOSRestoreImage in header
convertVZMacOSRestoreImage2Struct currently has a VZMacOSRestoreImage *
parameter, which causes the warning below when trying to build for an
older version of macOS with CGO_CFLAGS=-mmacosx-version-min=11.0.
This commit changes its type to `void *`. It is consistent with the
other `convert*` methods, and it fixes this warning:

 # github.com/Code-Hex/vz/v2
In file included from _cgo_export.c:4:
In file included from virtualization_arm64.go:10:
../../virtualization_arm64.h:68:61: warning: 'VZMacOSRestoreImage' is only available on macOS 12.0 or newer [-Wunguarded-availability-new]
/Library/Developer/CommandLineTools/SDKs/MacOSX.sdk/System/Library/Frameworks/Virtualization.framework/Headers/VZMacOSRestoreImage.h:30:12: note: 'VZMacOSRestoreImage' has been marked as being introduced in macOS 12.0 here, but the deployment target is macOS 11.0.0
../../virtualization_arm64.h:68:27: note: annotate 'convertVZMacOSRestoreImage2Struct' with an availability attribute to silence this warning