console: Don't return error for non-existing files
When using NewFileSerialPortAttachment(), the virtualization framework will
create the log file if it does not exist yet. If a directory in the path
is invalid, an error will be returned by the virtualization framework,
not an exception.This means we don't need to check this in Code-Hex/vz.

This commit removes the file existence check which was added in commit
117c12a.

This fixes https://github.com/Code-Hex/vz/issues/84