addtags: Fix `go vet` errors
With newer go versions, `go vet` reports:
```
cmd/addtags/main.go:19:25: non-constant format string in call to fmt.Fprintf
cmd/addtags/main.go:24:25: non-constant format string in call to fmt.Fprintf
```
This happens because of the use of [1:] to remove an unwanted `\n` at
the beginning of the string.
This commit switches to `Fprint` instead of `Fprintf` as there are no
substitutions to be made in the string.

Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>