socket: Make Accept() return after Close() is called
`Accept()` on a listener returned by `net.Listen()` will return an error
once the underlying listener is closed. Some packages in the ecosystem
rely on this behavior for orderly server shutdown.

Before this patch, `VirtioSocketDevice.Accept()` would never return once
the underlying listener was closed as `acceptch` was never closed. This
patch fixes this.