feat: add `VZVmnetNetworkDeviceAttachment` support (macOS 26.0)
`VZVmnetNetworkDeviceAttachment` is an API that creates vmnet devices on VMs added in macOS 26.

see: https://developer.apple.com/documentation/virtualization/vzvmnetnetworkdeviceattachment?language=objc

It does not require the `com.apple.vm.networking` entitlement nor root privileges.
`HostMode` and `SharedMode` are supported.
In order for multiple VMs to communicate with each other in `SharedMode`, they must be started in the same executable and the same `VmnetNetwork` must be passed to `NewVmnetNetworkDeviceAttachment()` to create an attachment.

This change adds:
- `vz.VmnetNetworkDeviceAttachment` represents `VZVmnetNetworkDeviceAttachment` in Go
- `vmnet` package to use vmnet APIs that added on macOS 26.0
  - `Return` represents `vmnet_return_t` as `error`
    - `ErrSuccess`, `ErrFailure`, ...
  - `Mode` represents `operating_modes_t`
    - `HostMode`, `SharedMode`
  - `NetworkConfiguration` represents `vmnet_network_configuration_t`
  - `Network` represents `vmnet_network_ref`

- `vz_test.TestVmnetSharedModeAllowsCommunicationBetweenMultipleVMs`
- `vz_test.TestVmnetSharedModeWithConfiguringIPv4`

Signed-off-by: Norio Nomura <norio.nomura@gmail.com>